Added updated training set without error
Error in original data set with cells containing multiple lines of data in one cell. This increased the word count to over 500 which was incorrect. Cleaned the original data set to no longer include this and adjust maximum word count in a sentence to be around 66, which is more realistic.